+++ src/main/java/com/zero/triptalk/plannerdetail/repository/PlannerDetailRepository.java
@@ -1,5 +1,6 @@
-package com.zero.triptalk.plannerdetail;
+package com.zero.triptalk.plannerdetail.repository;
 
+import com.zero.triptalk.plannerdetail.entity.PlannerDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
